Implementation for Completed table complete, bug fixes in Book.
You can not make two different deletes from the Completed table, one that completed gets rid of the book, the other that returns the book back to the reading list.

The reading list will only have one delete. To "move to finished reading", simply run a post call to /completed/id where the id is the one of the book from the booklist, and it will run the deletion on its own.

Attempted to begin implementing the update, but not working yet.