Book & Completed Database Entries Implemented with full CRUD and functional. (#14)
* API Call to Google Library now pushes book complete information to the backend database.

Now the booklist feature can directly fetch information from the database, instead of needing to rely on another API call. WE previously only stored information about ISBN, however not all books have this information. Books ae now stored with an id assigned by the database to uniquely identity each Book object.

Notable changes:
- Book model in frontend has all values set to datatype "string".
- API call passes in book data through the request body instead of parameters (and interpreted through the json bodyparser in app.mjs) to enure the preservation of the data.

* Resetting database information, preventing adding of test data.

* Adding some test data for usage in booklist.

* Implementation for Completed table complete, bug fixes in Book.

You can not make two different deletes from the Completed table, one that completed gets rid of the book, the other that returns the book back to the reading list.

The reading list will only have one delete. To "move to finished reading", simply run a post call to /completed/id where the id is the one of the book from the booklist, and it will run the deletion on its own.

Attempted to begin implementing the update, but not working yet.

* Update method/ call implemented and functional.

* Update Book.mjs

* Update Book.mjs